ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2